Fix build on musl-based environment

+++ src/server/reader_unix.cc
@@ -28,7 +28,7 @@
 namespace pixel_terrain::server {
     reader_unix::reader_unix(int fd) : fd(fd) {}
 
-    ssize_t reader_unix::fill_buffer(char *buf, size_t len, size_t off) {
+    long int reader_unix::fill_buffer(char *buf, size_t len, size_t off) {
         return read(fd, buf + off, len - off);
     }
